smoothers in calculators (#161)
* smoothers in calculators
