+++ tests/reference/berry_Chiral-ahc_iter-0000.dat
@@ -1,28 +1,28 @@
 #    Efermi                   x                  y                  z                  x                  y                  z        
   -5.000000e+00           0.000000e+00       0.000000e+00       0.000000e+00       0.000000e+00       0.000000e+00       0.000000e+00
-  -4.500000e+00           0.000000e+00       0.000000e+00       0.000000e+00       0.000000e+00       0.000000e+00       0.000000e+00
-  -4.000000e+00          -2.220446e-16       1.110223e-16       2.494786e+01      -2.220446e-16       1.110223e-16       2.494786e+01
-  -3.500000e+00           3.330669e-16      -4.440892e-16       8.337467e+01       3.330669e-16      -4.440892e-16       8.337467e+01
-  -3.000000e+00           1.554312e-15      -1.332268e-15       2.341319e+01       1.554312e-15      -1.332268e-15       2.341319e+01
-  -2.500000e+00           5.551115e-16      -2.220446e-16       1.456736e+02       5.551115e-16      -2.220446e-16       1.456736e+02
-  -2.000000e+00           8.881784e-16      -8.881784e-16       1.863804e+02       8.881784e-16      -8.881784e-16       1.863804e+02
-  -1.500000e+00          -3.252607e-19       2.168404e-19       2.086466e+01      -3.252607e-19       2.168404e-19       2.086466e+01
-  -1.000000e+00           3.155444e-30      -4.642775e-30       2.103813e+01       3.155444e-30      -4.642775e-30       2.103813e+01
-  -5.000000e-01           3.155444e-30      -4.642775e-30       2.105950e+01       3.155444e-30      -4.642775e-30       2.105950e+01
-   0.000000e+00           1.577722e-30      -1.992696e-30       1.372704e+02       1.577722e-30      -1.992696e-30       1.372704e+02
-   5.000000e-01          -8.881784e-16       8.881784e-16       1.854609e+02      -8.881784e-16       8.881784e-16       1.854609e+02
-   1.000000e+00          -2.428613e-17       2.081668e-17       1.395287e+02      -2.428613e-17       2.081668e-17       1.395287e+02
-   1.500000e+00           1.776357e-15      -1.332268e-15       1.154154e+02       1.776357e-15      -1.332268e-15       1.154154e+02
-   2.000000e+00          -6.938894e-17       2.775558e-17       1.611707e+02      -6.938894e-17       2.775558e-17       1.611707e+02
-   2.500000e+00           2.775558e-16      -3.330669e-16       7.312543e+01       2.775558e-16      -3.330669e-16       7.312543e+01
-   3.000000e+00           3.615612e-31      -9.696415e-31       1.999789e+01       3.615612e-31      -9.696415e-31       1.999789e+01
-   3.500000e+00          -1.301043e-18       1.734723e-18       5.147902e+00      -1.301043e-18       1.734723e-18       5.147902e+00
-   4.000000e+00           4.336809e-19      -4.336809e-19       1.703633e+00       4.336809e-19      -4.336809e-19       1.703633e+00
-   4.500000e+00           4.336809e-19      -4.336809e-19       7.587378e-01       4.336809e-19      -4.336809e-19       7.587378e-01
-   5.000000e+00           4.336809e-19      -4.336809e-19       2.375722e-01       4.336809e-19      -4.336809e-19       2.375722e-01
-   5.500000e+00           5.752111e-31      -1.035380e-30       2.136645e-02       5.752111e-31      -1.035380e-30       2.136645e-02
-   6.000000e+00           5.752111e-31      -1.035380e-30       2.136645e-02       5.752111e-31      -1.035380e-30       2.136645e-02
-   6.500000e+00           5.752111e-31      -1.035380e-30      -2.560547e-14       5.752111e-31      -1.035380e-30      -2.560547e-14
-   7.000000e+00           5.752111e-31      -1.035380e-30      -2.560547e-14       5.752111e-31      -1.035380e-30      -2.560547e-14
-   7.500000e+00           5.752111e-31      -1.035380e-30      -2.560547e-14       5.752111e-31      -1.035380e-30      -2.560547e-14
-   8.000000e+00           5.752111e-31      -1.035380e-30      -2.560547e-14       5.752111e-31      -1.035380e-30      -2.560547e-14
+  -4.500000e+00           0.000000e+00       0.000000e+00       0.000000e+00      -6.536794e-18       3.268397e-18       7.344426e-01
+  -4.000000e+00          -2.220446e-16       1.110223e-16       2.494786e+01      -1.991658e-16       9.141192e-17       2.593345e+01
+  -3.500000e+00           3.330669e-16      -4.440892e-16       8.337467e+01       3.526773e-16      -4.669680e-16       7.988942e+01
+  -3.000000e+00           1.554312e-15      -1.776357e-15       2.341319e+01       1.485676e-15      -1.691379e-15       2.877764e+01
+  -2.500000e+00           4.440892e-16      -2.220446e-16       1.456736e+02       4.898468e-16      -2.874125e-16       1.432727e+02
+  -2.000000e+00           8.881784e-16      -8.881784e-16       1.863804e+02       8.489449e-16      -8.424081e-16       1.803094e+02
+  -1.500000e+00          -4.336809e-19       4.336809e-19       2.086466e+01       2.573903e-17      -2.573903e-17       2.574240e+01
+  -1.000000e+00          -5.916457e-31      -6.080803e-31       2.103813e+01      -1.276718e-20       1.276718e-20       2.103366e+01
+  -5.000000e-01          -5.916457e-31      -6.080803e-31       2.105950e+01      -5.800340e-31      -6.545270e-31       2.448002e+01
+   0.000000e+00          -1.972152e-31      -2.185802e-30       1.372704e+02      -2.614718e-17       2.614718e-17       1.352679e+02
+   5.000000e-01          -8.881784e-16       8.881784e-16       1.854609e+02      -8.367012e-16       8.362926e-16       1.826900e+02
+   1.000000e+00          -2.775558e-17       1.387779e-17       1.395287e+02       2.580039e-20      -1.290020e-20       1.401710e+02
+   1.500000e+00           1.776357e-15      -1.332268e-15       1.154154e+02       1.668908e-15      -1.252600e-15       1.174722e+02
+   2.000000e+00          -6.938894e-17       2.775558e-17       1.611707e+02      -4.838095e-18      -2.617298e-17       1.572317e+02
+   2.500000e+00           2.775558e-16      -4.440892e-16       7.312543e+01       2.591710e-16      -4.171249e-16       7.415337e+01
+   3.000000e+00          -7.888609e-31      -4.601689e-31       1.999789e+01       8.132691e-18      -1.303529e-17       2.112475e+01
+   3.500000e+00          -1.301043e-18       1.301043e-18       5.147902e+00      -1.214864e-18       1.211672e-18       5.483676e+00
+   4.000000e+00           3.252607e-19      -4.336809e-19       1.703633e+00       2.773837e-19      -3.826122e-19       1.777212e+00
+   4.500000e+00           3.252607e-19      -4.336809e-19       7.587378e-01       3.252607e-19      -4.336809e-19       7.712121e-01
+   5.000000e+00           3.252607e-19      -4.336809e-19       2.375722e-01       3.156853e-19      -4.209137e-19       2.465500e-01
+   5.500000e+00          -8.874685e-31      -8.545993e-31       2.136645e-02       9.575382e-21      -1.276718e-20       2.773135e-02
+   6.000000e+00          -8.874685e-31      -8.545993e-31       2.136645e-02      -8.874685e-31      -8.545993e-31       2.073744e-02
+   6.500000e+00          -8.874685e-31      -8.545993e-31      -1.992976e-14      -8.874685e-31      -8.545993e-31       6.290091e-04
+   7.000000e+00          -8.874685e-31      -8.545993e-31      -1.992976e-14      -8.874685e-31      -8.545993e-31      -1.992976e-14
+   7.500000e+00          -8.874685e-31      -8.545993e-31      -1.992976e-14      -8.874685e-31      -8.545993e-31      -1.992976e-14
+   8.000000e+00          -8.874685e-31      -8.545993e-31      -1.992976e-14      -8.874685e-31      -8.545993e-31      -1.992976e-14
