+++ tests/reference/berry_Chiral-cumdos_iter-0000.dat
@@ -1,28 +1,28 @@
 #    Efermi                                               
   -5.000000e+00           0.000000e+00       0.000000e+00
-  -4.500000e+00           0.000000e+00       0.000000e+00
-  -4.000000e+00           2.250000e-02       2.250000e-02
-  -3.500000e+00           7.500000e-02       7.500000e-02
-  -3.000000e+00           2.250000e-01       2.250000e-01
-  -2.500000e+00           6.325000e-01       6.325000e-01
-  -2.000000e+00           8.650000e-01       8.650000e-01
-  -1.500000e+00           9.675000e-01       9.675000e-01
-  -1.000000e+00           9.825000e-01       9.825000e-01
-  -5.000000e-01           1.000000e+00       1.000000e+00
-   0.000000e+00           1.007500e+00       1.007500e+00
-   5.000000e-01           1.030000e+00       1.030000e+00
-   1.000000e+00           1.105000e+00       1.105000e+00
-   1.500000e+00           1.187500e+00       1.187500e+00
-   2.000000e+00           1.330000e+00       1.330000e+00
-   2.500000e+00           1.540000e+00       1.540000e+00
-   3.000000e+00           1.637500e+00       1.637500e+00
-   3.500000e+00           1.750000e+00       1.750000e+00
-   4.000000e+00           1.840000e+00       1.840000e+00
-   4.500000e+00           1.902500e+00       1.902500e+00
-   5.000000e+00           1.932500e+00       1.932500e+00
-   5.500000e+00           1.977500e+00       1.977500e+00
-   6.000000e+00           1.982500e+00       1.982500e+00
-   6.500000e+00           1.997500e+00       1.997500e+00
-   7.000000e+00           2.000000e+00       2.000000e+00
+  -4.500000e+00           0.000000e+00       6.623799e-04
+  -4.000000e+00           2.250000e-02       2.338317e-02
+  -3.500000e+00           7.500000e-02       7.787031e-02
+  -3.000000e+00           2.250000e-01       2.325806e-01
+  -2.500000e+00           6.325000e-01       6.273482e-01
+  -2.000000e+00           8.650000e-01       8.611729e-01
+  -1.500000e+00           9.675000e-01       9.649241e-01
+  -1.000000e+00           9.825000e-01       9.825736e-01
+  -5.000000e-01           1.000000e+00       9.997056e-01
+   0.000000e+00           1.007500e+00       1.007942e+00
+   5.000000e-01           1.030000e+00       1.031546e+00
+   1.000000e+00           1.105000e+00       1.105221e+00
+   1.500000e+00           1.187500e+00       1.189266e+00
+   2.000000e+00           1.330000e+00       1.331987e+00
+   2.500000e+00           1.540000e+00       1.536688e+00
+   3.000000e+00           1.637500e+00       1.637942e+00
+   3.500000e+00           1.750000e+00       1.749338e+00
+   4.000000e+00           1.840000e+00       1.839190e+00
+   4.500000e+00           1.902500e+00       1.901543e+00
+   5.000000e+00           1.932500e+00       1.932942e+00
+   5.500000e+00           1.977500e+00       1.976322e+00
+   6.000000e+00           1.982500e+00       1.982794e+00
+   6.500000e+00           1.997500e+00       1.997132e+00
+   7.000000e+00           2.000000e+00       1.999926e+00
    7.500000e+00           2.000000e+00       2.000000e+00
    8.000000e+00           2.000000e+00       2.000000e+00
