+++ tests/reference/berry_Chiral-dos_iter-0000.dat
@@ -1,28 +1,28 @@
 #    Efermi                                               
-  -5.000000e+00           0.000000e+00       0.000000e+00
-  -4.500000e+00           2.250000e-02       2.250000e-02
-  -4.000000e+00           7.500000e-02       7.500000e-02
-  -3.500000e+00           2.025000e-01       2.025000e-01
-  -3.000000e+00           5.575000e-01       5.575000e-01
-  -2.500000e+00           6.400000e-01       6.400000e-01
-  -2.000000e+00           3.350000e-01       3.350000e-01
-  -1.500000e+00           1.175000e-01       1.175000e-01
-  -1.000000e+00           3.250000e-02       3.250000e-02
-  -5.000000e-01           2.500000e-02       2.500000e-02
-   0.000000e+00           3.000000e-02       3.000000e-02
-   5.000000e-01           9.750000e-02       9.750000e-02
-   1.000000e+00           1.575000e-01       1.575000e-01
-   1.500000e+00           2.250000e-01       2.250000e-01
-   2.000000e+00           3.525000e-01       3.525000e-01
-   2.500000e+00           3.075000e-01       3.075000e-01
-   3.000000e+00           2.100000e-01       2.100000e-01
-   3.500000e+00           2.025000e-01       2.025000e-01
-   4.000000e+00           1.525000e-01       1.525000e-01
-   4.500000e+00           9.250000e-02       9.250000e-02
-   5.000000e+00           7.500000e-02       7.500000e-02
-   5.500000e+00           5.000000e-02       5.000000e-02
-   6.000000e+00           2.000000e-02       2.000000e-02
-   6.500000e+00           1.750000e-02       1.750000e-02
-   7.000000e+00           2.500000e-03       2.500000e-03
-   7.500000e+00           0.000000e+00       0.000000e+00
+  -5.000000e+00           0.000000e+00       6.824712e-04
+  -4.500000e+00           2.250000e-02       2.338317e-02
+  -4.000000e+00           7.500000e-02       7.720793e-02
+  -3.500000e+00           2.025000e-01       2.091974e-01
+  -3.000000e+00           5.575000e-01       5.494778e-01
+  -2.500000e+00           6.400000e-01       6.285923e-01
+  -2.000000e+00           3.350000e-01       3.375759e-01
+  -1.500000e+00           1.175000e-01       1.214007e-01
+  -1.000000e+00           3.250000e-02       3.478153e-02
+  -5.000000e-01           2.500000e-02       2.536799e-02
+   0.000000e+00           3.000000e-02       3.183994e-02
+   5.000000e-01           9.750000e-02       9.727921e-02
+   1.000000e+00           1.575000e-01       1.577208e-01
+   1.500000e+00           2.250000e-01       2.267663e-01
+   2.000000e+00           3.525000e-01       3.474218e-01
+   2.500000e+00           3.075000e-01       3.059544e-01
+   3.000000e+00           2.100000e-01       2.126495e-01
+   3.500000e+00           2.025000e-01       2.012488e-01
+   4.000000e+00           1.525000e-01       1.522056e-01
+   4.500000e+00           9.250000e-02       9.375116e-02
+   5.000000e+00           7.500000e-02       7.477921e-02
+   5.500000e+00           5.000000e-02       4.985280e-02
+   6.000000e+00           2.000000e-02       2.080958e-02
+   6.500000e+00           1.750000e-02       1.713201e-02
+   7.000000e+00           2.500000e-03       2.867989e-03
+   7.500000e+00           0.000000e+00       7.359776e-05
    8.000000e+00           0.000000e+00       0.000000e+00
