+++ tests/test_integrate.py
@@ -41,6 +41,7 @@ def check_integrate(parallel_serial):
             parameters_K={},
             specific_parameters={},
             use_symmetry=False,
+            smearEf=600.0,
             suffix="",
             suffix_ref="",
             mode="bin",
@@ -54,7 +55,7 @@ def check_integrate(parallel_serial):
             system,
             grid=grid,
             Efermi=Efermi,
-            smearEf=600.0,
+            smearEf=smearEf,
             #            omega = omega,
             quantities=quantities,
             user_quantities=user_quantities,
@@ -667,6 +668,7 @@ def test_Chiral(check_integrate, system_Chiral_left, compare_energyresult, quant
         quantities_Chiral,
         fout_name="berry_Chiral",
         Efermi=Efermi_Chiral,
+        smearEf=1200,
         comparer=compare_energyresult,
         use_symmetry=True,
         additional_parameters={'external_terms': False},
@@ -684,6 +686,7 @@ def test_Chiral_txt(check_integrate, system_Chiral_left, compare_energyresult, q
         Efermi=Efermi_Chiral,
         comparer=compare_energyresult,
         mode="txt",
+        smearEf=1200,
         use_symmetry=True,
         precision=-1e-5,
         additional_parameters={'external_terms': False},
@@ -700,6 +703,7 @@ def test_Chiral_tetra(check_integrate, system_Chiral_left, compare_energyresult,
         fout_name="berry_Chiral_tetra",
         Efermi=Efermi_Chiral,
         comparer=compare_energyresult,
+        smearEf=1200,
         use_symmetry=True,
         additional_parameters={
             'external_terms': False,
