+++ tests/test_run.py
@@ -6,6 +6,7 @@ import pytest
 
 import wannierberri as wberri
 from wannierberri import calculators as calc
+from wannierberri.smoother import FermiDiracSmoother
 from wannierberri.__result import EnergyResult
 
 from common import OUTPUT_DIR
@@ -99,10 +100,12 @@ calculators_Te = {
     'berry_dipole': calc.static.BerryDipole_FermiSea,
 }
 
+
+smoother_Chiral = FermiDiracSmoother(Efermi_Chiral, T_Kelvin=1200, maxdE=8)
 calculators_Chiral = {
-    'conductivity_ohmic': calc.static.Ohmic(Efermi=Efermi_Chiral),
-    'berry_dipole': calc.static.BerryDipole_FermiSea(Efermi=Efermi_Chiral, kwargs_formula={"external_terms": False}),
-    'ahc': calc.static.AHC(Efermi=Efermi_Chiral, kwargs_formula={"external_terms": False})
+    'conductivity_ohmic': calc.static.Ohmic(Efermi=Efermi_Chiral,smoother=smoother_Chiral),
+    'berry_dipole': calc.static.BerryDipole_FermiSea(Efermi=Efermi_Chiral, kwargs_formula={"external_terms": False},smoother=smoother_Chiral),
+    'ahc': calc.static.AHC(Efermi=Efermi_Chiral, kwargs_formula={"external_terms": False},smoother=smoother_Chiral)
 }
 
 
@@ -150,7 +153,7 @@ def test_Fe(check_run, system_Fe_W90, compare_any_result, compare_fermisurfer):
             0,
             fout_name_ref="kubo_Fe_W90",
             suffix_ref=quant,
-            precision=1e-8,
+            precision=-1e-8,
             result_type=EnergyResult)
 
     extra_precision = {'berry': 1e-6}
@@ -263,7 +266,7 @@ def test_Fe_pickle_Klist(check_run, system_Fe_W90, compare_any_result):
     )
 
 
-def test_GaAs(check_run, system_GaAs_W90, compare_any_result, compare_fermisurfer):
+def test_GaAs(check_run, system_GaAs_W90, compare_any_result):
     param = {'Efermi': Efermi_GaAs}
     calculators = {k: v(**param) for k, v in calculators_GaAs.items()}
 
@@ -280,7 +283,7 @@ def test_GaAs(check_run, system_GaAs_W90, compare_any_result, compare_fermisurfe
     )  # This is a low precision for the nonabelian thing, not sure if it does not indicate a problem, or is a gauge-dependent thing
 
 
-def test_Chiral_left(check_run, system_Chiral_left, compare_any_result, compare_fermisurfer):
+def test_Chiral_left(check_run, system_Chiral_left, compare_any_result, compare_energyresult):
     grid_param = {'NK': [10, 10, 4], 'NKFFT': [5, 5, 2]}
     check_run(
         system_Chiral_left,
@@ -296,8 +299,18 @@ def test_Chiral_left(check_run, system_Chiral_left, compare_any_result, compare_
         extra_precision={"Morb": -1e-6},
     )
 
+    for quant in calculators_Chiral.keys():#["conductivity_ohmic", "berry_dipole", "ahc"]:
+        compare_energyresult(
+                fout_name="berry_Chiral",
+                suffix=quant+"-left-run",
+                adpt_num_iter=0,
+                suffix_ref=quant,
+                mode="txt",
+                compare_smooth=True,
+                precision=-1e-8)
 
-def test_Chiral_leftTR(check_run, system_Chiral_left, system_Chiral_left_TR, compare_any_result, compare_fermisurfer):
+
+def test_Chiral_leftTR(check_run, system_Chiral_left, system_Chiral_left_TR, compare_any_result):
     "check that for time-reversed model the ohmic conductivity is the same, but the AHC is opposite"
     grid_param = {'NK': [10, 10, 4], 'NKFFT': [5, 5, 2]}
     results = [
@@ -323,7 +336,7 @@ def test_Chiral_leftTR(check_run, system_Chiral_left, system_Chiral_left_TR, com
         assert data1 == pytest.approx(sign * data2, abs=precision), key
 
 
-def test_Chiral_right(check_run, system_Chiral_left, system_Chiral_right, compare_any_result, compare_fermisurfer):
+def test_Chiral_right(check_run, system_Chiral_left, system_Chiral_right, compare_any_result):
     "check that for flipped chirality the ohmic conductivity is the same, but the Berry dipole is opposite"
     grid_param = {'NK': [10, 10, 4], 'NKFFT': [5, 5, 2]}
     results = [
