+++ wannierberri/__integrate.py
@@ -15,7 +15,8 @@ import numpy as np
 from collections import defaultdict
 from copy import copy
 
-from .__utility import VoidSmoother, TAU_UNIT
+from .__utility import TAU_UNIT
+from .smoother import VoidSmoother
 from . import __result as result
 from . import fermiocean as fermiocean
 from . import __kubo as kubo
