+++ wannierberri/__main.py
@@ -15,7 +15,7 @@
 import functools
 from . import __version__
 from .__evaluate import evaluate_K
-from .__utility import getSmoother
+from .smoother import getSmoother
 from . import __integrate
 from . import __tabulate
 from .__path import Path
